+++ tests/StreamTest.php
@@ -1418,10 +1418,10 @@ final class StreamTest extends TestCase
         
         self::assertTrue($result->found());
         self::assertSame([1, 2, 3, 4], $result->get());
-        self::assertSame([[1, 2, 3, 4]], $result->toArray());
-        self::assertSame([[1, 2, 3, 4]], $result->toArrayAssoc());
+        self::assertSame([1, 2, 3, 4], $result->toArray());
+        self::assertSame([1, 2, 3, 4], $result->toArrayAssoc());
         self::assertSame('[1,2,3,4]', $result->toJson());
-        self::assertSame('[[1,2,3,4]]', $result->toJsonAssoc());
+        self::assertSame('[1,2,3,4]', $result->toJsonAssoc());
         self::assertSame([0, [1, 2, 3, 4]], $result->tuple());
         
         $counter = 0;