+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All important changes to `fiisoft-jackdaw-stream` will be documented in this fil
 - added method StreamApi::collect
 - added method StreamApi::aggregate
 - method Result::toString() accepts param `string $separator = ','`
+- changed methods in Result: toJson, toJsonAssoc, toArray, toArrayAssoc - they accept default parameters as StreamApi and work different for array-results
+- identical methods from StreamApi and Result moved to ResultCaster 
 - class FiiSoft\Jackdaw\Collector\Collectors\Collect renamed to CollectIn
 - composer.lock added to .gitignore